removed topic and moved entire logic to lucene generator
